device_wrapper.c: 6: yes: MISRA.SHIFT.RANGE.2012
Fixed MISRA rule 12.2 violations by casting shift operands to uint32_t
- rule 12.2, MISRA.SHIFT.RANGE.2012: Right operand of shift operation
is out of range - greater or equal to the essential type size of left
operand or is negative, 3 violations fixed, IDs 686, 687, 688, tfa
enforced

+++ plat/ti/k3low/common/drivers/pm/device/device_wrapper.c
@@ -11,8 +11,8 @@
 #include <device_wrapper.h>
 #include <string.h>
 
-#define POWER_STATE_ON	(0 << 30)
-#define POWER_STATE_OFF (1 << 30)
+#define POWER_STATE_ON	((uint32_t)0U << 30)
+#define POWER_STATE_OFF ((uint32_t)1U << 30)
 
 int32_t scmi_handler_device_state_set_on(uint32_t dev_id)
 {
